requirements.txt
cupy==7.5.0
easydict==1.9
imageio==2.8.0
ipdb==0.13.3
matplotlib==3.2.2
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.2
pypng==0.0.20
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.1.0
scs==2.1.2
six==1.14.0
sklearn==0.0
tensorboardX==2.0
tqdm==4.54.0
